requirements.txt
Pillow==6.0.0
PyYAML==5.1
